requirements.txt
feedparser==6.0.11
newspaper3k==0.2.8
weasyprint==65.1
python-dotenv==1.1.0
readability-lxml==0.8.1
